Divide monolithic translations.{cpp,h} into separate files
