+++ src/martialarts.cpp
@@ -25,6 +25,7 @@
 #include "item_factory.h"
 #include "itype.h"
 #include "json.h"
+#include "localized_comparator.h"
 #include "map.h"
 #include "output.h"
 #include "pimpl.h"
