+++ src/translation_manager.h
@@ -6,6 +6,7 @@
 
 #include <memory>
 #include <unordered_set>
+#include <vector>
 
 class TranslationManagerInterface
 {
