+++ tests/catacharset_test.cpp
@@ -9,6 +9,7 @@
 
 #include "catacharset.h"
 #include "cata_catch.h"
+#include "localized_comparator.h"
 #include "translations.h"
 
 TEST_CASE( "utf8_width", "[catacharset]" )
